Add Specific Retirement for Toucan Protocol (#11)
Replace `redeemAuto` with `redeemAuto2` to save on gas, as well as implement check
to ensure that the `retire` function is only called on TCO2s actually redeemed.

Add methods to selectively choose and retire TCO2s using the aggregator. This accounts
for any redemption fees incurred so that the desired number of tons will be retired.

Update the `getNeededBuyAmount` function across all helpers to allow consistent switching
between factoring in a redemption fee or just the AMM fees.